pyproject.toml
[project]
name = "crawl4ai-mcp"
version = "0.1.0"
description = "MCP server for integrating web crawling and RAG into AI agents and AI coding assistants"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai==0.6.2",
    "mcp==1.7.1",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "openai==1.71.0",
    "anthropic>=0.40.0",
    "requests>=2.31.0",
    "dotenv==0.9.9",
    "sentence-transformers>=4.1.0",
    "neo4j>=5.28.1",
]
